The updating of the cross position is shifted by -yB because the trajectory was using the helix method to compute the position at the cylinder radius and it was not accounting for the beam spot.  Actually this computation was not needed because the KF already reports the trajectory.  Both these issues have been fixed.
The hemisphere was wrong in the making of the surfaces, as it took the middle of the cluster strip which for BMT C gives the wrong hemisphere for events in a particular phi range.  This has been fixed by using the afore mentioned method to get the point from the seed and get the hemisphere from there.