fix: remove test-scoped dependency `org.jlab.jnp:jnp-hipo4` from reconstruction tests
This was strange. If I scope `jnp-hipo4` to `test` in the POM file, as
suggested by `mvn dependency:analyze`, I would get build errors in very
specific methods, for example, `ECCommon.readStripsHipo`:
- my IDE (Eclipse JDT LSP) correctly sees `DataEvent` as `org.jlab.io.base.DataEvent`
- the Maven build (`maven-compiler-plugin`) would get confused and think
  that `DataEvent` comes from somewhere in `jnp-hipo4`, even though this
  package was not being imported

With the help of @raffaelladevita we found that `jnp-hipo4` dependency
could actually be removed from all of the tests, so we can avoid the
issue altogether.