build: remove repo mirrors in favor of POM `<repository>`
To be more explicit which repositories are actually being used, let's
get rid of the mirrors. Only `freehep-repo-public` mirror was _actually_
being used by the POMs:
```xml
<mirror>
  <id>new-freehep-repo-public</id>
  <mirrorOf>freehep-repo-public</mirrorOf>
  <name>freehep-repo-public</name>
  <url>https://clasweb.jlab.org/clas12maven</url>
</mirror>
```
Notice the URL is slightly different from the one we use in the POMs.
Still, even after this change, some dependencies are being downloaded
from
```
https://clasweb.jlab.org/clas12maven
```
and not
```
https://clasweb.jlab.org/.clas12maven
```
I suspect this is because some dependencies' POM files specify
repository URLs too, namely:
- https://clasweb.jlab.org/.clas12maven/org/jlab/groot/4.0.5/
- https://clasweb.jlab.org/.jhep/maven/org/jlab/jnp/jnp-hipo/2.0-SNAPSHOT/